add support for cancelation of builds with snapshot dependencies
In the current state, the plugin assumes that if a target build in the queue
it will always run the latest revision and will not try to cancel it.

This is not always the case for builds with snapshot dependency. Based
on the teamcity configuration, a snapshot dependencies can enforce the
revision of a queued build if one of the dependencies has already started

For example:
Triggering a build A with a snapshot dependency B configured with "Enforce revisions synchronization".
As soon as B leave the queue, it will enforce the revision on the build A, if the VCS root is the same for both builds

The plugin now will ensure the revision of the target build is correct
before skipping it and will with an opt-in option cancel any non reused dependencies.